Refactor Tooltip into its own separate component
