+++ src/components/NormalRightRow/index.jsx
@@ -1,8 +1,7 @@
 import React from 'react';
 import { shape, string } from 'prop-types';
 import Typography from '@material-ui/core/Typography';
-import Tooltip from '@material-ui/core/Tooltip';
-import Warning from '@material-ui/icons/Info';
+import Tooltip from '../Tooltip';
 
 function NormalRightRow({ schema, classes }) {
   /**
@@ -21,23 +20,6 @@ function NormalRightRow({ schema, classes }) {
   const keywords = Object.keys(schema).filter(
     key => !skipKeywords.includes(key)
   );
-  /**
-   * Generate tooltip descriptions to provide further information
-   * regarding the given keywords.
-   * (only keywords defined as complex object types will need
-   *  tooltip descriptions)
-   */
-  const tooltipDescriptions = {
-    additionalItems: 'Additional items must match a sub-schema',
-    additionalProperties: 'Additional properties must match a sub-schema',
-    dependencies:
-      'The schema of the object may change based on the presence of certain special properties',
-    propertyNames: 'Names of properties must follow a specified convention',
-    patternProperties:
-      'Property names or values should match the specified pattern',
-  };
-  const createTooltipTitle = key =>
-    `${tooltipDescriptions[key]}. See the JSON-schema source for details.`;
 
   return (
     <div className={`${classes.row} ${classes.rightRow}`}>
@@ -53,12 +35,7 @@ function NormalRightRow({ schema, classes }) {
               className={classes.line}>
               {typeof schema[keyword] === 'object' &&
               !Array.isArray(schema[keyword]) ? (
-                <Tooltip title={createTooltipTitle(keyword)} arrow>
-                  <Typography component="span" variant="subtitle2">
-                    {`${keyword}: `}
-                    <Warning fontSize="inherit" color="inherit" />
-                  </Typography>
-                </Tooltip>
+                <Tooltip keyword={keyword} />
               ) : (
                 `${keyword}: ${schema[keyword]}`
               )}
